feature: allow the use of source and canname
To use source in the template, use {Source}. However I do not recommend this
because the source can change as new devices are added or the bus is reset.

The preferred method would be to use canname. This is a unique identifier to
the device and will not change. To use canname in the template, use {canName}